src/requirements.txt
Django>=2.1.7
djangorestframework==3.9.2
django-crispy-forms==1.7.2
django-model-utils==3.1.2
Pillow==5.0.0
PyYAML>=4.2b1
Markdown==2.6.11
django-admin-rangefilter==0.3.14
psycopg2==2.7.3.2
flake8==3.7.7